refactor(parser): extract types
